1_foundations/requirements.txt
requests
python-dotenv
gradio
pypdf
openai
openai-agents
langchain
langchain-community
langchain_openai
tiktoken
chromadb
matplotlib